ml-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning libraries
torch==2.1.0
transformers==4.35.2
pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.4

# HTTP client for API calls
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Redis for caching
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Model serving
onnxruntime==1.16.3  # For faster inference